+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Keyed list reconciliation for reactive runtime children
+
 ## [0.1.0] - 2026-03-20
 
 ### Added